performacne_scenario: Don't Omit Hosts
Due to a bump of `ansible-lint` in commit
7c41c797d8ed1c2da71c3cc8489d1618d8c348fd, we were required to omitt
variable if they are not defined when using them in hosts section.
This has caused the playbook to fail for OpenShift based deployments
because they do not require to provide certain values.

Since the issue 'syntax-check' can not be ignored by `ansible-lint`, we
will exclude 'playbooks/packet_gen/trex/performance_scenario.yml' from
being linted.

Change-Id: Ibec02cdc44aa30bd7920b863bc056b6473ca6212