ansible: do no change clouds.yaml directory
"mode: 0755" changes the permssions of an already existing directory
when they do not match. That was causing that the permissions of the
already created home directory (e.g. /home/stack) are changed and
causing in turn that the public key authentication against
undercloud-0 stops working (requires 0775 for home directory in
RHEL). Instead of using 0775, 0700 is preferred because clouds.yaml
file may contain sensitive data (e.g. passwords).

Change-Id: I2117037fc1cf309d36e38f6bbe56da56752f9cad