ansible: Update `switches_config` Role
Updating the `switches_config` role to do a hacky test of Junos configuration database lock.
By default, the role will attempt to check if the database is not locked three times in intervals of 900 seconds.

Removing incorrect attempt definitions for switches_config role.

Updating `ansible.cfg` to maintain a connection for a long time.

Change-Id: I2b89dd7309aa4b40389c59fae0e7ccdb3bdc2f03