Add nohz_full and rcu_nocbs to vm grub
nohz_full
nohz_full gives us reduced ticks when there is only 1 task on given cpu

rcu_nocbs
The Read-Copy-Update (RCU) system is a lockless mechanism for mutual
exclusion inside the kernel. As a consequence of performing RCU operations,
call-backs are sometimes queued on CPUs to be performed at a future moment
when removing memory is safe.

Change-Id: I95868cf9dbc0e379b06a8f2f39b7342d402d3a48